assets folder issue, handled
