+++ package.json
@@ -14,8 +14,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-postcss": "^4.0.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -47,12 +45,16 @@
     "@babel/core": "^7.23.6",
     "@babel/preset-env": "^7.23.6",
     "@babel/preset-react": "^7.23.3",
+    "@rollup/plugin-alias": "^5.1.0",
     "@svgr/rollup": "^8.1.0",
     "rollup": "^2.79.1",
     "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-copy": "^3.5.0",
     "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-svg": "^2.0.0",
     "rollup-plugin-url": "^3.0.1"
   },
   "peerDependencies": {
