+++ src/PokerPlayingCards.jsx
@@ -1,17 +1,17 @@
-import { Back } from './Back.jsx';
-import { Joker } from './Joker.jsx';
-import { Ace } from './Ace.jsx';
-import { Two } from './Two.jsx';
-import { Three } from './Three.jsx';
-import { Four } from './Four.jsx';
-import { Five } from './Five.jsx';
-import { Six } from './Six.jsx';
-import { Seven } from './Seven.jsx';
-import { Eight } from './Eight.jsx';
-import { Nine } from './Nine.jsx';
-import { Ten } from './Ten.jsx';
-import { Jack} from './Jack.jsx';
-import { Queen } from './Queen.jsx';
-import { King } from './King.jsx';
+import { Back } from './components/Back.jsx';
+import { Joker } from './components/Joker.jsx';
+import { Ace } from './components/Ace.jsx';
+import { Two } from './components/Two.jsx';
+import { Three } from './components/Three.jsx';
+import { Four } from './components/Four.jsx';
+import { Five } from './components/Five.jsx';
+import { Six } from './components/Six.jsx';
+import { Seven } from './components/Seven.jsx';
+import { Eight } from './components/Eight.jsx';
+import { Nine } from './components/Nine.jsx';
+import { Ten } from './components/Ten.jsx';
+import { Jack} from './components/Jack.jsx';
+import { Queen } from './components/Queen.jsx';
+import { King } from './components/King.jsx';
 
 export { Back, Joker, Ace, Two, Three, Four, Five, Six, Seven, Eight, Nine, Ten, Jack, Queen, King};
