+++ src/components/Queen.jsx
