+++ src/components/Six.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import styles from './assets/css/styles.css';
+
 import * as Constants from './Constants.jsx';
 import { BaseCard } from './BaseCard.jsx';
